Add files via upload
covered the following edge case which may arise during NE-AE comms: the NE may send a message containing the end of the current flow+metadata and the beginning of the next one like this: message= "dport":80}{"packets": . The AE, in its current state, would simply check the last character of the message to see if it is equal to the delimiter ("}") in order to determine whether or not it should perform inference and send back a confirmation signal. In this case, the AE would completely overlook the end of the current message. After this commit, the AE now searches the entire message for the delimiter, isolates the appropriate part of the message, sends a confirmation signal,  performs inference, and then isolates the start of the next message before it continues into the next iteration of the read loop.